# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kolibri AI OS V2.1",
    "version": "2.1.0"
  },
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Capabilities",
        "operationId": "capabilities_v1_capabilities_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/tools": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Tools",
        "operationId": "tools_v1_tools_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/runtime/summary": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Runtime Summary",
        "operationId": "runtime_summary_v1_runtime_summary_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/runtime/dags": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Runtime Dags",
        "operationId": "runtime_dags_v1_runtime_dags_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/runtime/actors": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Runtime Actors",
        "operationId": "runtime_actors_v1_runtime_actors_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
...
    "/v1/nodes/register": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Register Node",
        "operationId": "register_node_v1_nodes_register_post",
        "parameters": [
          {
            "name": "x-kolibri-node-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kolibri-Node-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeRegistration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "KolibriNodeToken": []
          }
        ]
      }
    },
    "/v1/nodes/{node_id}/heartbeat": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Heartbeat Node",
        "operationId": "heartbeat_node_v1_nodes__node_id__heartbeat_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "x-kolibri-node-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kolibri-Node-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeHeartbeat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "KolibriNodeToken": []
          }
        ]
      }
    },
    "/v1/fleet/nodes": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Fleet Nodes",
        "operationId": "fleet_nodes_v1_fleet_nodes_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/tasks": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Create Task",
        "operationId": "create_task_v1_tasks_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_v1_tasks_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Get Task",
        "operationId": "get_task_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/tasks/lease": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Lease Task",
        "operationId": "lease_task_v1_tasks_lease_post",
        "parameters": [
          {
            "name": "x-kolibri-node-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kolibri-Node-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "KolibriNodeToken": []
          }
        ]
      }
    },
    "/v1/tasks/{task_id}/heartbeat": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Task Heartbeat",
        "operationId": "task_heartbeat_v1_tasks__task_id__heartbeat_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "x-kolibri-node-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kolibri-Node-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseHeartbeat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "KolibriNodeToken": []
          }
        ]
      }
    },
    "/v1/tasks/{task_id}/artifacts": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Upload Task Artifact",
        "operationId": "upload_task_artifact_v1_tasks__task_id__artifacts_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "x-kolibri-node-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kolibri-Node-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_task_artifact_v1_tasks__task_id__artifacts_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "KolibriNodeToken": []
          }
        ]
      }
    },
    "/v1/tasks/{task_id}/complete": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Complete Task",
        "operationId": "complete_task_v1_tasks__task_id__complete_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "x-kolibri-node-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kolibri-Node-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskComplete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "KolibriNodeToken": []
          }
        ]
      }
    },
    "/v1/tasks/{task_id}/fail": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Fail Task",
        "operationId": "fail_task_v1_tasks__task_id__fail_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "x-kolibri-node-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Kolibri-Node-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskFail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "KolibriNodeToken": []
          }
        ]
      }
    },
    "/v1/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Task Events",
        "operationId": "task_events_v1_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeHeartbeat": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "online"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          },
          "active_attempts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Attempts"
          }
        },
        "type": "object",
        "title": "NodeHeartbeat"
      },
      "NodeRegistration": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "NodeRegistration"
      },
...
      "TaskCreate": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "health_probe"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": "Kolibri task"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input"
          },
          "required_capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Capabilities"
          },
          "required_artifacts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Artifacts"
          },
          "max_attempts": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "default": 3
          }
        },
        "type": "object",
        "title": "TaskCreate"
      },
      "TaskFail": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "TaskFail"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sk-kolibri API key or signed Kolibri session",
        "description": "OpenAI-compatible bearer authentication."
      },
      "KolibriSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kolibri_session",
        "description": "Secure HttpOnly browser session cookie."
      },
      "KolibriSessionHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Session"
      },
      "KolibriNodeToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Node-Token"
      }
    }
  },
...
}